Move useGetDatasetItems hook
